// db/db.d.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export interface Users {
  created_at: Generated<Date>;
  email: Generated<string | null>;
  first_name: Generated<string | null>;
  id: Generated<number>;
  is_active: Generated<number | null>;
  last_name: Generated<string | null>;
  password: Generated<string | null>;
  roles: Generated<string | null>;
  updated_at: Generated<Date | null>;
}

export interface DB {
  users: Users;
}